fix barrier roll error
