+++ src/module.api/stats.controller.ts
@@ -8,6 +8,7 @@ import BigNumber from 'bignumber.js'
 import { PriceTickerMapper } from '@src/module.model/price.ticker'
 import { MasternodeStats, MasternodeStatsMapper } from '@src/module.model/masternode.stats'
 import { BlockchainInfo } from '@defichain/jellyfish-api-core/dist/category/blockchain'
+import { getBlockSubsidy } from '@src/module.api/subsidy'
 
 @Controller('/stats')
 export class StatsController {
@@ -198,26 +199,6 @@ export function getEmission (eunosHeight: number, height: number): StatsData['em
     burned: burned.toNumber(),
     total: total.toNumber()
   }
-}
-
-export function getBlockSubsidy (eunosHeight: number, height: number): BigNumber {
-  let blockSubsidy = new BigNumber(405.04)
-
-  if (height >= eunosHeight) {
-    const reductionAmount = new BigNumber(0.01658) // 1.658%
-    const reductions = Math.floor((height - eunosHeight) / 32690) // Two weeks
-
-    for (let i = reductions; i > 0; i--) {
-      const amount = reductionAmount.times(blockSubsidy)
-      if (amount.lte(0.00001)) {
-        return new BigNumber(0)
-      }
-
-      blockSubsidy = blockSubsidy.minus(amount)
-    }
-  }
-
-  return blockSubsidy
 }
 
 function requireValue<T> (value: T | undefined, name: string): T {
